+++ app/src/main/java/eu/kanade/tachiyomi/ui/home/HomeScreen.kt
@@ -42,6 +42,7 @@ import eu.kanade.tachiyomi.ui.library.LibraryTab
 import eu.kanade.tachiyomi.ui.local.LocalTab
 import eu.kanade.tachiyomi.ui.manga.MangaScreen
 import eu.kanade.tachiyomi.ui.more.MoreTab
+import eu.kanade.tachiyomi.ui.vault.ImportRequestsTab
 import eu.kanade.tachiyomi.ui.vault.VaultTab
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.collectLatest
@@ -73,6 +74,7 @@ object HomeScreen : Screen() {
         LibraryTab,
         LocalTab,
         VaultTab,
+        ImportRequestsTab,
         BrowseTab,
         MoreTab,
     )
@@ -154,6 +156,7 @@ object HomeScreen : Screen() {
                             is Tab.Library -> LibraryTab
                             Tab.Local -> LocalTab
                             Tab.Vault -> VaultTab
+                            Tab.ImportRequests -> ImportRequestsTab
                             is Tab.Browse -> {
                                 if (it.toExtensions) {
                                     BrowseTab.showExtension()
@@ -281,6 +284,7 @@ object HomeScreen : Screen() {
         data class Library(val mangaIdToOpen: Long? = null) : Tab
         data object Local : Tab
         data object Vault : Tab
+        data object ImportRequests : Tab
         data class Browse(val toExtensions: Boolean = false) : Tab
         data class More(val toDownloads: Boolean) : Tab
     }